python-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
umbral @ git+https://github.com/nucypher/pyUmbral.git@v0.3.0
ipfshttpclient==0.7.0
web3==7.3.0
python-dotenv==1.0.1
typer==0.12.5
rich==13.9.3
pydantic==2.9.2

